HinghwaDict/requirements.txt
django
demjson
pyjwt
django-cors-headers
xlrd
cos-python-sdk-v5
apscheduler
django_apscheduler
django-notifications-hq
